+++ src/openforms/submissions/logic/rules.py
@@ -1,11 +1,11 @@
 from dataclasses import dataclass
-from typing import Callable, Dict, Iterable, Iterator, List, Optional
+from typing import Callable, Iterable, Iterator, List, Optional
 
 import elasticapm
 from json_logic import jsonLogic
 
 from openforms.forms.constants import LogicActionTypes
-from openforms.forms.models import FormLogic, FormStep, FormVariable
+from openforms.forms.models import FormLogic, FormStep
 
 from ..models import Submission, SubmissionStep
 from .actions import ActionOperation, compile_action_operation
@@ -161,22 +161,3 @@ def iter_evaluate_rules(
                 else:
                     operation = compile_action_operation(action)
                     yield operation
-
-
-def get_targeted_components(
-    rule: FormLogic,
-    components_map: dict,
-    all_variables: Dict[str, FormVariable],
-    initial_data: dict,
-) -> List[dict]:
-    targeted_components = []
-    for action in rule.actions:
-        action = ActionOperation.from_action(action)
-        action_log_data = action.get_action_log_data(
-            component_map=components_map,
-            all_variables=all_variables,
-            initial_data=initial_data,
-        )
-        targeted_components.append(action_log_data)
-
-    return targeted_components
